gimhana/requirements.txt
openai-whisper>=20231117
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.1.0
numpy>=1.24.0
